+++ front/components/assistant/conversation/input_bar/InputBarUsageBanner.tsx
@@ -1,4 +1,4 @@
-import { UsageUpgradeButton } from "@app/components/credits/RequestUpgradeButton";
+import { UsageUpgradeButton } from "@app/components/credits/UsageUpgradeButton";
 import { useAuth } from "@app/lib/auth/AuthContext";
 import { useWorkspaceUsageStatus } from "@app/lib/swr/user";
 import type { LightWorkspaceType } from "@app/types/user";
